web functionality done. debuging

+++ core/src/server/web_impl/dto/StatusDto.hpp
@@ -30,7 +30,7 @@ class StatusDto: public oatpp::data::mapping::type::Object {
     DTO_INIT(StatusDto, Object)
 
     DTO_FIELD(String, message, "message");
-    DTO_FIELD(Int32, code, "error-code");
+    DTO_FIELD(Int64, code, "error-code");
 };
 
 #include OATPP_CODEGEN_END(DTO)
